edge/traffic/requirements.txt
absl-py==0.4.1
aiohttp==3.4.2
alembic==1.0.10
astor==0.7.1
async-timeout==3.0.0
attrs==19.1.0
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2019.3.9
cffi==1.10.0
chardet==3.0.4
Click==7.0
cupy==4.1.0
cycler==0.10.0
Cython==0.26
decorator==4.4.0
Django==1.9.5
dlib==19.17.0
easydict==1.9
entrypoints==0.2.3
fastrlock==0.4
Flask==1.0.2
Flask-Cors==3.0.3
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SocketIO==2.8.5
Flask-SQLAlchemy==2.3.2
futures==3.1.1
gast==0.2.0
gevent==1.2.2
gevent-websocket==0.10.1
greenlet==0.4.12
grpcio==1.12.1
h5py==2.9.0
html5lib==0.9999999
idna==2.8
idna-ssl==1.1.0
imageio==2.4.1
imageio-ffmpeg==0.3.0
imgaug==0.2.9
imutils==0.5.2
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==1.1.0
jedi==0.10.2
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
kafka-python==1.4.6
Keras==2.1.5
kiwisolver==1.0.1
lmdb==0.94
mahotas==1.4.5
Mako==1.0.9
Markdown==2.6.9
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.7.4
mmdnn==0.2.5
moviepy==1.0.0
msgpack==0.6.1
multidict==4.4.0
mysqlclient==1.4.2.post1
nbconvert==5.2.1
nbformat==4.4.0
networkx==2.3
notebook==5.0.0
numpy==1.16.2
olefile==0.44
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
pika==1.0.1
Pillow==6.0.0
plotly==3.8.0
proglog==0.1.9
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.5.2
pycparser==2.18
pycrypto==2.6.1
Pygments==2.2.0
PyMySQL==0.7.9
pyparsing==2.2.0
pyrsistent==0.14.11
python-dateutil==2.6.1
python-editor==1.0.4
python-engineio==1.3.1
python-socketio==1.7.4
pytorch2keras==0.1.16
pytz==2017.2
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.2.1
redis==3.2.1
requests==2.21.0
retrying==1.3.3
salt==2019.2.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
Shapely==1.6.4.post2
simplegeneric==0.8.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.1.13
tensorboard==1.10.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
terminado==0.6
testpath==0.3
tflearn==0.3.2
Theano==1.0.4
torch==1.1.0
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.3.0
tornado==4.5.3
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.2
visdom==0.1.8.8
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.2
widgetsnbextension==3.0.2
yarl==1.2.6
